fix(calculator): let Calculator correctly use module name for checks
This fixes a bug in MROCalculator._is_original_class(), which checks if
a Jedi Name is an originally defined class in a script by check if the
full name of the class starts with the full name of the module.
This check was previously using the script.get_context().full_name,
however get_context() is not returning the context of the script, but the
context of the last position of the script and can give wrong result if
the last position is in a class declaration etc.
Now `script.get_context().full_name` changes to
`script.get_context().module_name` to correctly get the module name.