test(complex_mro): add a new test example
This patch adds a new test example which checks if the language server
can correctly resolve the MRO list for complex inheritance situation.
All the class declarations are in the same module and this example is
taken from the wiki page for C3 linearisation. The original author takes
the credit and please refer to the wiki page for more details.