fix(calculator): use name.goto() to fetch full context Jedi Name
This patch corrects the bug that a Jedi Name for class may not have
full content like `full_name` when using the returned results of
methods like get_names(). The solution is to call name.goto() and
retrieve the first return (which should be available) to get the full
context of the target Jedi Name.