refactor(calculator): move content_cache from analyser to calculator
This patch moves the content_cache, which is responsible
for caching the lines of the scripts, even the unsaved parts
from MROAnalyser to MROCalculator.

Now the MROAnalyser is purely a wrapper class over
MROCalculator to provide API for the language server
functions, and the MROCalculator is the core class for  the
MRO relevant calculations.

This patch also fixes some small bugs to make MROCalculator
supporting MRO calculations for class declarations
referencing across multiple files.