test(class_def_in_multi_files): add new test examples
This patch adds a set of new test examples to check if the
language server can deal with a complex MRO list where
all the relevant class declarations are distributed in
different files.

All the files are in the `class_def_in_multi_files` sub
package and the examples are taken from the wiki page
of C3 linearisation. The credit belongs to the original author.