refactor(package.json): remove unused contributes
This patch removes the three unused `contributes.configuration` in the
package.json setting file, as currently the language server is not
following these settings and these settings are not representative for
the functionalities of the language server.

A later feature should add appropriate setting options back in the
package.json to give user the control of behaviour of the MRO language
server.