chore(ci): update CI to only run on ubuntu
This patch updates the CI process so that it will only test over Ubuntu
platform for now, because of issues of failing to construct a CI process
to install the project and run the tests for MacOC and Windows.
The CI for these two platforms will be added in a following update asap.